// frontend/src/app/core/api/model/promotion-stats-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PromotionStatsResponse { 
    promotionName?: string;
    academicYearLabel?: string;
    totalSubjects?: number;
    totalCoefficient?: number;
    globalHourlyVolume?: number;
    promotionType?: string;
    trainingLabel?: string;
}