// frontend/src/app/core/api/model/promotion-subject-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PromotionSubjectRequest { 
    promotionId: number;
    subjectId: number;
    academicPeriodId?: number;
    coefficient: number;
}